+++ src/main/java/cartographer/cache/SurfaceCacheTile.java
@@ -372,7 +372,7 @@ public final class SurfaceCacheTile implements CachedSurfaceTileView {
 
     private static long checkedTileStart(int coordinate, String axis) {
         try {
-            return Math.multiplyExact((long) coordinate, (long) MapChunk.SIZE);
+            return Math.multiplyExact(coordinate, (long) MapChunk.SIZE);
         } catch (ArithmeticException exception) {
             throw new IllegalArgumentException("mapchunk " + axis + " coordinate overflows", exception);
         }
@@ -383,7 +383,7 @@ public final class SurfaceCacheTile implements CachedSurfaceTileView {
         if (remaining <= 0 || remaining > Integer.MAX_VALUE) {
             throw new IllegalArgumentException("derived Surface tile dimension is invalid");
         }
-        return (int) Math.min((long) MapChunk.SIZE, remaining);
+        return (int) Math.min(MapChunk.SIZE, remaining);
     }
 
     record Geometry(int width, int height) {
